hw10

+++ src/main/java/edu/hw10/Task1TestClass.java
@@ -1,17 +1,18 @@
 package edu.hw10;
+import edu.hw10.Task1Annotations.*;
 
 public class Task1TestClass {
 
     public final boolean ok;
 
-    public final long n;
+    public final int n;
 
-    public Task1TestClass(boolean ok1, long k) {
+    public Task1TestClass(boolean ok1, int k) {
         ok = ok1;
         n = k;
     }
 
-    public static Task1TestClass create(boolean ok1, long k) {
+    public static Task1TestClass create(boolean ok1, int k) {
         return new Task1TestClass(ok1, k);
     }
 }
